Move event and config duktape extensions out of ovms_duktape to their individual modules (which we can now do as we have split off ovms_duktape from ovms_script/ovms_command)
