+++ vehicle/OVMS.V3/components/ovms_script/srcduk/ovms_duktape.cpp
@@ -577,177 +577,6 @@ static duk_ret_t DukOvmsAssert(duk_context *ctx)
   return 0;
   }
 
-static duk_ret_t DukOvmsRaiseEvent(duk_context *ctx)
-  {
-  const char *event = duk_to_string(ctx,0);
-  uint32_t delay_ms = duk_is_number(ctx,1) ? duk_to_uint32(ctx,1) : 0;
-
-  if (event != NULL)
-    {
-    MyEvents.SignalEvent(event, NULL, (size_t)0, delay_ms);
-    }
-  return 0;  /* no return value */
-  }
-
-static duk_ret_t DukOvmsConfigParams(duk_context *ctx)
-  {
-  if (!MyConfig.ismounted()) return 0;
-
-  duk_idx_t arr_idx = duk_push_array(ctx);
-  int count = 0;
-  for (ConfigMap::iterator it=MyConfig.m_map.begin(); it!=MyConfig.m_map.end(); ++it)
-    {
-    duk_push_string(ctx, it->first.c_str());
-    duk_put_prop_index(ctx, arr_idx, count++);
-    }
-
-  return 1;
-  }
-
-static duk_ret_t DukOvmsConfigInstances(duk_context *ctx)
-  {
-  const char *param = duk_to_string(ctx,0);
-
-  if (!MyConfig.ismounted()) return 0;
-
-  OvmsConfigParam *p = MyConfig.CachedParam(param);
-  if (p)
-    {
-    if (! p->Readable()) return 0;  // Parameter is protected, and not readable
-
-    duk_idx_t arr_idx = duk_push_array(ctx);
-    int count = 0;
-    for (ConfigParamMap::iterator it=p->m_map.begin(); it!=p->m_map.end(); ++it)
-      {
-      duk_push_string(ctx, it->first.c_str());
-      duk_put_prop_index(ctx, arr_idx, count++);
-      }
-    return 1;
-    }
-
-  return 0;
-  }
-
-static duk_ret_t DukOvmsConfigGetValues(duk_context *ctx)
-  {
-  const char *param = duk_to_string(ctx,0);
-  std::string prefix = duk_opt_string(ctx,1,"");
-
-  if (!MyConfig.ismounted()) return 0;
-
-  OvmsConfigParam *p = MyConfig.CachedParam(param);
-  if (p)
-    {
-    if (! p->Readable()) return 0;  // Parameter is protected, and not readable
-
-    duk_idx_t obj_idx = duk_push_object(ctx);
-    for (ConfigParamMap::iterator it=p->m_map.lower_bound(prefix); it!=p->m_map.end(); ++it)
-      {
-      if (!startsWith(it->first, prefix)) break;
-      duk_push_string(ctx, it->second.c_str());
-      duk_put_prop_string(ctx, obj_idx, it->first.substr(prefix.length()).c_str());
-      }
-    return 1;
-    }
-
-  return 0;
-  }
-
-static duk_ret_t DukOvmsConfigSetValues(duk_context *ctx)
-  {
-  const char *param = duk_to_string(ctx,0);
-  std::string prefix = duk_opt_string(ctx,1,"");
-  duk_require_object(ctx,2);
-
-  if (!duk_is_object(ctx,2)) return 0;
-  if (!MyConfig.ismounted()) return 0;
-
-  OvmsConfigParam *p = MyConfig.CachedParam(param);
-  if (p)
-    {
-    if (! p->Writable()) return 0;  // Parameter is not writeable
-
-    ConfigParamMap pmap = p->m_map;
-    std::string key, val;
-    duk_enum(ctx, 2, 0);
-    while (duk_next(ctx, -1, true))
-      {
-      key = duk_to_string(ctx, -2);
-      val = duk_to_string(ctx, -1);
-      duk_pop_2(ctx);
-      pmap[prefix+key] = val;
-      }
-    duk_pop(ctx);
-    p->m_map.clear();
-    p->m_map = std::move(pmap);
-    p->Save();
-    }
-
-  return 0;
-  }
-
-static duk_ret_t DukOvmsConfigGet(duk_context *ctx)
-  {
-  const char *param = duk_to_string(ctx,0);
-  const char *instance = duk_to_string(ctx,1);
-  const char *defvalue = duk_to_string(ctx,2);
-
-  if (!MyConfig.ismounted()) return 0;
-  OvmsConfigParam *p = MyConfig.CachedParam(param);
-  if (p)
-    {
-    if (! p->Readable()) return 0;  // Parameter is protected, and not readable
-    if (p->IsDefined(instance))
-      {
-      std::string v = p->GetValue(instance);
-      duk_push_string(ctx, v.c_str());
-      }
-    else
-      {
-      duk_push_string(ctx, defvalue);
-      }
-    return 1;
-    }
-  else
-    {
-    return 0;
-    }
-  return 0;
-  }
-
-static duk_ret_t DukOvmsConfigSet(duk_context *ctx)
-  {
-  const char *param = duk_to_string(ctx,0);
-  const char *instance = duk_to_string(ctx,1);
-  const char *value = duk_to_string(ctx,2);
-
-  if (!MyConfig.ismounted()) return 0;
-
-  OvmsConfigParam *p = MyConfig.CachedParam(param);
-  if (p)
-    {
-    if (! p->Writable()) return 0;  // Parameter is not writeable
-    p->SetValue(instance, value);
-    }
-  return 0;
-  }
-
-static duk_ret_t DukOvmsConfigDelete(duk_context *ctx)
-  {
-  const char *param = duk_to_string(ctx,0);
-  const char *instance = duk_to_string(ctx,1);
-
-  if (!MyConfig.ismounted()) return 0;
-
-  OvmsConfigParam *p = MyConfig.CachedParam(param);
-  if (p)
-    {
-    if (! p->Writable()) return 0;  // Parameter is not writeable
-    p->DeleteInstance(instance);
-    }
-  return 0;
-  }
-
 ////////////////////////////////////////////////////////////////////////////////
 // DuktapeObject
 
@@ -2006,21 +1835,8 @@ OvmsDuktape::OvmsDuktape()
   RegisterDuktapeModule(mod_json_js_start, mod_json_js_end - mod_json_js_start, "JSON");
 
   // Register standard functions
-  DuktapeObjectRegistration* dto;
   RegisterDuktapeFunction(DukOvmsPrint, 1, "print");
   RegisterDuktapeFunction(DukOvmsAssert, 2, "assert");
-  dto = new DuktapeObjectRegistration("OvmsEvents");
-  dto->RegisterDuktapeFunction(DukOvmsRaiseEvent, 2, "Raise");
-  RegisterDuktapeObject(dto);
-  dto = new DuktapeObjectRegistration("OvmsConfig");
-  dto->RegisterDuktapeFunction(DukOvmsConfigParams, 0, "Params");
-  dto->RegisterDuktapeFunction(DukOvmsConfigInstances, 1, "Instances");
-  dto->RegisterDuktapeFunction(DukOvmsConfigGet, 3, "Get");
-  dto->RegisterDuktapeFunction(DukOvmsConfigSet, 3, "Set");
-  dto->RegisterDuktapeFunction(DukOvmsConfigDelete, 2, "Delete");
-  dto->RegisterDuktapeFunction(DukOvmsConfigGetValues, 2, "GetValues");
-  dto->RegisterDuktapeFunction(DukOvmsConfigSetValues, 3, "SetValues");
-  MyDuktape.RegisterDuktapeObject(dto);
   DuktapeObjectRegistration* dt_http = new DuktapeObjectRegistration("HTTP");
   dt_http->RegisterDuktapeFunction(DuktapeHTTPRequest::Create, 1, "request"); // legacy
   dt_http->RegisterDuktapeFunction(DuktapeHTTPRequest::Create, 1, "Request");
